gemini_apikey_chatbot/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
google-generativeai
python-multipart 